Make test_process pass again, instead of hang.
Calling simuate in addReader meant that it was possible for a child process to
terminate, raise SIGCHLD, and get prematurely reaped before
process.Process.__init__ finished and the pid to be registered properly for
reaping.

Does anyone know why simulate was being called here?  The tests still pass
without it...


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@9682 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb