merge -r12229:HEAD $tslybr/trial-tearDownIssue .
fix trial so that failures in setUp and tearDown propagate to their 
test methods. Before, in setUp and tearDown were considered autonomous.
Warner, however pointed out that they are really like factored-out 
beginnings and endings of each method, and therefore should count as 
part of those methods.

This commit also contains requisite changes to the reporters to produce
sane output


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@12231 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb