resolve issue 781
cannot make wait() reentrant without using threads, instead raise exception

wait keeps a counter to make sure that it is only waiting on one deferred
at a time and is not reentered. If it is reentered, it raises a 
util.WaitIsNotReentrantError. there is no code in twisted that currently 
relies upon this behavior of wait, so I am opting to say "Don't do that" 
for the moment, and coming up with the nasty "real" answer if a valid 
use-case surfaces.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@12402 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb