So just kidding about the other __path__ fix. This one might work better.
Strategy:
instead of doing a recursive walk over the directory structure, do it one
directory at a time.
Given a package (module object), look at its __path__, and iterate
over the directories in the list. For each file in the directory, check
if it matches the test case pattern. For each subdir, check if it's a package,
and if so, try to import it. If it imported, recurse, again given the actual
package's module's __path__.

Problems:
- Breaks some trial tests
- Displays import errors from packages like twisted.internet.serialport
which cannot be imported.
- No new tests.



git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/branches/foom@14851 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb