<pre>An error has occurred.</pre> if t.w.html.output() fails - resolves issue 1337
r872 removed the traceback from the <pre></pre>, leaving it 
empty, but comes with no explaination and failed to update 
the docstring to indicate the change in behavior.  Most likely
this was deemed a security fix, to avoid revealing source lines,
filesystem details, and the like.  Nothing in Twisted uses this,
and other parts of twisted.web gleefully include complete
traceback text in content delivered to clients, however I will
not re-duplicate that behavior here.  Instead, change output()
to log the exception and return a non-descript error message.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@15270 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb