Retool locateChild() a bit in File.
The old implementation of locateChild would sometimes return a resource other than the resource cooresponding to the given path, such as:
 - When a directory is located, return a DirectoryLister or the resource for the index file
 - If a processor is in place, return whatever the processor returns
 - If a resource doesn't exist,
   - return a sibling matching ignoredExts (which really seems like a misnomer, by the way)
   - or return None

There are some problems here which presently trip up DAV, but represent a couple of deep issues.

First, a directory (collection) resource is not the same resource as the index file or the DirectoryLister which is used to render the collection during a GET request.  For example, a PUT of COPY of the collection isn't the same at all as PUT or COPY on these stand-in resources, they may have different metadata, etc.  As such, returning a different resource is inappropriate.

If a resource doesn't exist, we still want to return a resource object, not None, because we may be needing to create that resource.  Resource.exists() is an appropriate way to determine whether the resource you get back actually exists or not.

What this commit does is it does not return a stand-in or None in these cases, and instead returns the requested resource.  In order to preserve the present functionality provided by stand-in resources for collections, we instantiate the stand-in in render() and return the result of it's render() method.

We still have some magic going on in locateChild() in the case of processors.  This seems OK insofar as the processed resource is the correct resource for the requested URI.  In the DAV case, we don't allow processed resources, so things remain kosher.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@15523 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb