Handle either unicode or byte strings as avatarIds in the realm
This kind of sucks.  I think avatarIds should probably always be unicode.

However, previously this code was broken as the IRC server always tried 
to log in with a unicode string and the realm always demanded a byte
string.

This also fixes a bug in error handling during login, which previously
assumed all errors were either authentication failures or duplicate
login attempts, and silently discarded any other kind of failure.



git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@15762 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb