Merge of Issue 654 branch.
Author: glyph
Reviewer: exarkun

This refactors process support for win32, adds support for
twisted.internet.stdio to windows, and adds an IProducer and working IConsumer
implementation.

Notably, as of this merge, win32 users can spawn worker subprocesses with
spawnProcess that use Twisted and StandardIO and they will work as expected.
Considering that this is a suggested idiom for Twisted programs it is good that
it's now portable.



git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@15834 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb