Apply mwh's epytext cleanup patch
Author: mwh
Reviewer: exarkun
Fixes #1943

This fixes various errors in the epytext markup in twisted.conch and twisted.application.
It also makes changes to correct markup to make it more useful and so that it renders
in a more appealing way.  Documentation which was not using epytext markup at all has
also been converted to use it.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@17656 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb