Merge twistd-plugins-1922-3
Author: radix
Reviewer: exarkun
Fixes #1922

The first step towards a mktapless world: take three.

"Oh, ye so fiercely tended,
 Ye little seeds of hate!
I bent above your growing
 Early and noon and late,
Yet are ye drooped and pitiful,--
 I cannot rear ye straight!"

twistd now can accept subcommands equivalent to mktap's;
it will look up the IServiceMaker plugin for the given subcommand
and use it in precisely the same way that mktap does, except that it
immediately applies the service to an Application and starts it.

That is to say,

"twistd web --path ."



git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@17992 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb