Import minidom within a try/except to cause this not to break platforms without minidom.
Author: Stephen Thorne
Reviewer: jml 
Fixes: #2076

twisted.python.versions will use an xml parser to read twisted/.svn/entries.
This would break on platforms such as pypy where no xml parser is available
(pyexpat and minidom are both non functional on this platform). In order to
allow pypy to work with twisted out of svn, this fix where we catch the
ImportError from importing minidom is required.



git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@18086 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb