Merge epollreactor-1953
Author: therve, exarkun 
Reviewer: PenguinOfDoom, Jerub, jml
Fixes #1953

Add a level-triggered epoll-based reactor.

Also re-order the removal and the shutdown of sockets from the reactor
in the tcp code shared between reactors in order to avoid possibly passing
a no-longer valid file descriptor to a control API such as epoll_ctl() or
kevent().


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@18475 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb