Use `LoggedSuite`  instead of `startKeepingErrors` to trap logged errors  * Author: jml  * Reviewers: Jerub, exarkun  * Fixes #2173
`startKeepingErrors` has been deprecated for a while, but Trial was still
using it to catch errors that were logged during a test run, but outside 
of any particular test case. 

This branch creates a new kind of `TestSuite` called `LoggedSuite`, which
monitors the log and adds any logged errors to the test result object.



git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@18764 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb