Merge ctrl+backslash-2371+2123-5
Author: exarkun
Reviewer: jknight, radix, therve
Fixes #2371, #2123, #2378

Remove the long-since obsolete test helper SignalMixin, which set up
and tore down special signal handling state.  For some time, trial has
taken care of this detail. (#2378)

Change the process file descriptor "connection lost" code to reverse
the setNonBlocking operation done during initialization, leaving the
file descriptor in the blocking state (what it probably was before it
was given to the process code).  This works around a bug with PTYs on
OS X where the last write made to a non-blocking PTY might be lost if
the process exits before the data is read from the other end of the
PTY.  This also corrects a (perhaps unreasonable) assumption that
blocking writes may be made to stdout after twisted.internet.stdio
has been used.  (#2371)

Change ConsoleManhole so that it waits for connectionLost notification
before stopping the reactor.  This avoids the possibility that the
reactor will stop before its transport has had all of its output
flushed.  Combined with the fix for #2371, this fixes #2123.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@19640 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb