Merge web2-anon-auth-2460
Author: dreid
Reviewer: therve
Fixes #2460

Add support for anonymous authentication with the HTTPAuthResource.
HTTPAuthResource will attempt to authenticate with ANONYMOUS credentials
if no Authorization header is given.  If the portal has an AllowAnonymousAccess 
checker registered with it then this will succeed.  

If at any time an HTTPError is raised with an UNAUTHORIZED status code 
HTTPAuthResource will add WWW-Authenticate headers based on it's current
configuration (if the WWW-Authenticate header is absent)



git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@19764 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb