Useful tracebacks when running Trial tests against pyunit `TestResult`s
 * Author: jml
 * Reviewer: exarkun
 * Fixes #2495

Adds a new method `getTracebackObject` to `Failure`, which always returns 
an object that can be passed to `traceback.extract_tb` and includes all the
information the `Failure` has on the relevant stack trace.

Uses this new method to provide `sys.exc_info`-like information to pyunit
`TestResult` objects via `PyUnitResultAdapter`.



git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@19771 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb