Cleanup empty callback lists in EventDispatcher.
Author: ralphm
Reviewer: therve, exarkun
Fixes #1652, #2476.

EventDispatcher did not remove empty callback lists in the event callback
registration. This causes events to be matched to queries that will not
trigger any callback being called, causing significant slowdown when
many one-time callbacks are registered.

This change makes sure empty callback lists are cleaned out, and also optimizes
the internal bookkeeping to not have to use sorting on every callback
registration. Instead the observers are kept in different dictionaries (query
-> callback list) for each priority level. Then, upon event dispatch, the
observers priorities are sorted and then each associated dictionary is
processed in that order. This assumes the number of used priority levels is
very low to be efficient.

This remerge also fixes a regression introduced in r19669 that made multiple
observers for the same query fail (#2476).


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@19807 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb