Merge plugin-collision-2339
Author: exarkun
Reviewer: therve
Fixes #2339

Add a function, twisted.plugin.pluginPackagePaths, which returns a list of
additional directories to search for plugin modules for a particular plugin
package.  Change twisted.plugins.__init__ to use this function instead of
looping over sys.path directly.

Unlike the loop previously in twisted.plugins.__init__, pluginPackagePaths
will not include any directory which is a Python package.  This prevents
plugins from being discovered from packages which Python may not actually
be able to import and limits plugin discovery to those installed on the
version of the package which is in use by an application.  It also allows
a site installation to co-exist with another installation (for example,
a personal user installation) without interfering with each other.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@20920 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb