Merge addcallbacks-2849
Author: peaker (mostly), exarkun
Reviewer: jerub, therve, exarkun
Fixes #2849

Fix Deferred.addCallback for the case where it is called on a Deferred from
a callback on that same Deferred.  The behavior is changed so that the new
callback is added to the end of the callback chain and only called after
the executing callback returns at the earliest.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@21529 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb