Correct the behavior of `proxyForInterface` with respect to inheritance.
Author: glyph

Reviewer: therve

Fixes #2901

This corrects two bugs in `proxyForInterface`.  The first is that subclasses of types generated for `proxyForInterface` were not providing convincing facimilies of the methods that they allegedly implemented; you could not define a proxy class `Foo` and a subclass `Bar`, and have `Bar` call `Foo.method(self)`.  Now you can.

The other is that proxyies for sub-interfaces were incorrectly restricting access to the methods of their superclasses.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@21987 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb