Merge personal-servers-2681
Author: exarkun
Reviewer: therve
Fixes #2681

Factor creation of the distrib/personal server factory into a separate function
which can easily be unit tested, add a test for it, and fix it so that it uses
a PBServerFactory (a class which exists) instead of BrokerFactory (a class which
does not exist).

This fixes "twistd web --personal".


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@22538 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb