Merge test-process-alias-3110
Author: therve
Reviewer: radix
Fixes #3110

This gets rid of a script file from the filesystem and instead dynamically
generates the script at test-execution time. This avoids a problem where
easy_install would not maintain the executable bit of the script; now
that the test dynamically generates the script, it can explicitly mark it
as executable.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@23001 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb