Merge process-ended-3116-2
Author: jml, exarkun
Reviewer: therve
Fixes: #3116

Add a unit test for the handling of exceptions from
`IProcessProtocol.processEnded` implementations.  Make
sure the error is logged and reactor state associated
with it is still cleaned up.

Also fix the name of the Windows Event reactor in the
reactor builder harness so that it is possible for that
reactor to be tested on platforms where it works.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@24056 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb