Merge concurrent-trial_temp-2338
Author: exarkun
Reviewer: therve
Fixes: #2338

Add a lock to the `_trial_temp` directory so that different test
processes don't try to use the same directory at the same time.
If no temporary directory is specified, a different default will
silently be selected if the preferred default is found to be in
use.  If a temporary directory is specified and found to be in
use, an error will be reported.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@24325 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb