Merge firsterror-display-3298
Author: exarkun, itamar
Reviewer: thijs, radix
Fixes: #3298

Change the `__str__` and `__repr__` of `FirstError` so that they include better
and more complete information about the underlying failure (and therefore the
real cause of the exception).  This means that when these errors are logged, the
log will now include information about the exception which caused the `FirstError`.

The deprecated tuple-like interface is also removed.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@25200 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb