Merge timezones-are-impossible-3551
Author: exarkun
Reviewer: mwhudson
Fixes: #3551

Change the unit tests for `twisted.conch.ls.lsLine` so to set the current timezone for
the process so that the results are more easily predictable.  This fixes the unit tests
in timezones where the values used in the test fall on a different local day than the
tests expected.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@25489 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb