Merge minidom-helpers-3600
Author: exarkun
Reviewer: therve
Fixes: #3600

Change the existing `twisted.web.domhelpers` test suite so that it also runs
using `xml.dom.minidom` as the DOM implementation, in addition to still running
while using `twisted.web.microdom` as the DOM implementation.  Make the necessary
changes to `twisted.web.domhelpers` so that the suite passes with both implementations.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@25917 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb