Apply better-exception.patch
Author: exarkun
Reviewer: therve
Fixes: #3692

Change the text of the `VerifyError` raised when an attempt is made to associate
a private key with a `PrivateCertificate` which does not match that certificate's
public key.  The text now states explicitly what the mismatch was, rather than
just making a general statement about an improperly signed certificate.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@26690 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb