Merge disable-connection-sharing-3498-2
Author: z3p
Reviewer: exarkun, therve, glyph
Fixes: #3498
Refs #3483, #3497, #716

Remove the Conch SSH connection sharing functionality.  This feature is not well
tested and has numerous subtle bugs which can interfer with the normal operation
of the Conch SSH client.

The Conch command line arguments which supported configuring this functionality
have been removed (supplying them is now an error) has has the module primarily
responsible for the implementation.

It is expected that this feature will be re-introduced at some point, without
the problems afflicting this implementation.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@26756 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb