Merge mailbox-order-3812
Author: exarkun
Reviewer: glyph
Fixes: #3812

Fix an intermittent failure in the message ordering of the maildir-based
mailbox implementation.  Previously messages delivered within the same
second had a small chance to be delivered in such a way which that they
would appear to have been delivered in the opposite order as they really
were.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@26821 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb