Merge httpauth-method-not-allowed-3932
Author: therve, exarkun
Reviewer: therve
Fixes: #3932

Change `DeferredResource` so that it uses the request object to call the
`render` method of whatever resource its deferred eventually fires with.
This makes handling of the result (by return or exception) of such render
methods consistent with normal handling.  Previously only strings and
`NOT_DONE_YET` were correctly handled by `DeferredResource`, nothing else
(for example raising `UnsupportedMethod` was not handling, resulting in
a 500 response instead of the correct 405 or 501).


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@27595 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb