Merge test-signalkill-4186
Author: exarkun
Reviewer: radix
Fixes: #4186

Make several of the tests for child process functionality more robust by
having them account for the possibility that a write from the child they
interact with might be delivered in multiple chunks to the process
protocol.

This should fix a handful of intermittent `ProcessAlreadyExited` errors
in tests on `PosixProcessTestCase`.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@27892 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb