Merge failure-unicode-4501
Author: exarkun
Reviewer: thijs
Fixes: #4501

Fix the error handling in `Failure.printTraceback` in the "brief"
detail mode so that exceptions which raise an exception when being
converted to a string do not cause `printTraceback` itself to raise
an exception.

Re-merge includes the last revision from the branch which was
accidentally omitted in the first merge.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@29704 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb