Merge zippath-child-insecure-4535
Author: exarkun
Reviewer: therve
Fixes: #4535

Change the repr string of `twisted.python.zippath.ZipPath` instances to more
faithfully represent the underlying path.  Previously, incorrect platform-specific
processing would be applied to the path string before returning it from the
repr method, causing it to be misleading in the presence of ".." (and possible
other) path components.

Also document the potentially surprising behavior of `ZipPath.child` with
respect to special (or rather, not special) path components such as "..".


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@29797 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb