Merge web-template-4939 - incorporate the good parts of Nevow into Twisted.
Author: habnabit, glyph, lewq, jesstess, jerub, exarkun

Reviewer: glyph, idnar

Fixes: #4939

This change pulls in exarkun's latest and greatest page rendering model (hence
his well-deserved inclusion on the authors list, even though he didn't make
commits to this ''particular'' branch) from `nevow.page.Element`, with several
modifications to simplify it and reduce the amount of code in it.  Also included
is a new narrative tutorial document explaining how to use basic templating, and
also that you ''DO NOT NEED THIS'' to generate HTML in a Twisted program; you
just might want it because it's easy and fun.

This should allow us to much more gracefully and correctly generate HTML within
Twisted in the few places we already do, and perhaps also allow us to more
easily provide new browser-based, simple management tools for Twisted servers at
some point in the future.



git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@31357 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb