Merge win32-gtk-win32events-4862
Author: exarkun
Reviewer: itamarst
Fixes: #4862

Add `IReactorWin32Events` to the select, iocp, and Gtk reactors on Windows
(if the third-party dependencies are available) by having each of those
reactors run a `Win32Reactor` in a secondary thread and delegating event-
related methods to that reactor.

Significantly, this means all reactors on Windows now support serial ports.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@32109 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb