Merge wrapped-log-prefix-5062-3
Author: therve, retenodus, exarkun
Reviewer: zseil, exarkun, glyph
Fixes: #5062

Change transport implementations to use the protocol's ILoggingContext
implementation (when it is available) to construct the value for the
system key in the logging context, instead of always using the protocol
class's name directly.

This improves the behavior when using endpoints or `twisted.protocols.policies`
wrappers, since the wrapper protocol can include information about the wrapped
protocol in the log prefix.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@32728 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb